# src/docs/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "ShortLink API",
		"version": "1.0.0",
		"license": {
			"name": "MIT",
			"identifier": "MIT"
		}
	},
	"security": [
		{
			"bearerAuth": []
		}
	],
	"paths": {
		"/urls/{code}": {
			"get": {
				"tags": ["urls"],
				"summary": "Resolve a short code to its target url",
				"security": [],
				"parameters": [
					{
						"name": "code",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Resolved",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Result"
										},
										{
											"type": "object",
											"properties": {
												"result": {
													"$ref": "#/components/schemas/Url"
												}
											}
										}
									]
								}
							}
						}
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					}
				}
			},
			"patch": {
				"tags": ["urls"],
				"summary": "Repoint an owned short link",
				"parameters": [
					{
						"name": "code",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["url"],
								"properties": {
									"url": {
										"type": "string",
										"format": "uri"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Result"
										},
										{
											"type": "object",
											"properties": {
												"result": {
													"$ref": "#/components/schemas/Url"
												}
											}
										}
									]
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"422": {
						"$ref": "#/components/responses/Invalid"
					}
				}
			},
			"delete": {
				"tags": ["urls"],
				"summary": "Delete an owned short link",
				"parameters": [
					{
						"name": "code",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Removed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					}
				}
			}
		},
		"/urls": {
			"get": {
				"tags": ["urls"],
				"summary": "List the short links owned by the caller",
				"responses": {
					"200": {
						"description": "Listed",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Result"
										},
										{
											"type": "object",
											"properties": {
												"result": {
													"type": "array",
													"items": {
														"$ref": "#/components/schemas/Url"
													}
												}
											}
										}
									]
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					}
				}
			},
			"post": {
				"tags": ["urls"],
				"summary": "Shorten a url",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["url"],
								"properties": {
									"url": {
										"type": "string",
										"format": "uri"
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Shortened",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Result"
										},
										{
											"type": "object",
											"properties": {
												"result": {
													"$ref": "#/components/schemas/Url"
												}
											}
										}
									]
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"422": {
						"$ref": "#/components/responses/Invalid"
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"tags": ["auth"],
				"summary": "Authenticate and receive a JWT",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Credentials"
							}
						},
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/Credentials"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Result"
										},
										{
											"type": "object",
											"properties": {
												"result": {
													"$ref": "#/components/schemas/Auth"
												}
											}
										}
									]
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Invalid"
					},
					"422": {
						"$ref": "#/components/responses/Invalid"
					}
				}
			}
		},
		"/auth/register": {
			"post": {
				"tags": ["auth"],
				"summary": "Register an account and receive a JWT",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Credentials"
							}
						},
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/Credentials"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Registered",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Result"
										},
										{
											"type": "object",
											"properties": {
												"result": {
													"$ref": "#/components/schemas/Auth"
												}
											}
										}
									]
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Invalid"
					},
					"422": {
						"$ref": "#/components/responses/Invalid"
					}
				}
			}
		}
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"responses": {
			"Unauthorized": {
				"description": "Missing, malformed or expired token",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Result"
						}
					}
				}
			},
			"NotFound": {
				"description": "Short link is not found",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Result"
						}
					}
				}
			},
			"Invalid": {
				"description": "Missing or invalid fields",
				"content": {
					"application/json": {
						"schema": {
							"allOf": [
								{
									"$ref": "#/components/schemas/Result"
								},
								{
									"type": "object",
									"properties": {
										"result": {
											"$ref": "#/components/schemas/FieldErrors"
										}
									}
								}
							]
						}
					}
				}
			}
		},
		"schemas": {
			"Url": {
				"type": "object",
				"properties": {
					"url": {
						"type": "string",
						"format": "uri"
					},
					"encoded": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"Result": {
				"type": "object",
				"required": ["success", "message", "result"],
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {}
				}
			},
			"FieldErrors": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				}
			},
			"Credentials": {
				"type": "object",
				"required": ["email", "password"],
				"properties": {
					"email": {
						"type": "string",
						"format": "email"
					},
					"password": {
						"type": "string"
					}
				}
			},
			"Auth": {
				"type": "object",
				"properties": {
					"user": {
						"type": "object",
						"properties": {
							"email": {
								"type": "string",
								"format": "email"
							}
						}
					},
					"token": {
						"type": "string"
					}
				}
			}
		}
	},
	"tags": [
		{
			"name": "urls",
			"description": "Short link management"
		},
		{
			"name": "auth",
			"description": "Registration and login"
		}
	]
}